requirements.txt
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
google-api-core==2.6.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-translate==3.7.1
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==21.3
pdf2image==1.16.0
Pillow==9.0.1
pkg_resources==0.0.0
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.7
pytesseract==0.3.9
python-dotenv==0.19.2
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
toml==0.10.2
urllib3==1.26.8
